Add files via upload
This project aims to study the use of linear programming (LP) in finance and explore in which way linear programming can be brought up and used for Cost of Capital 
with the emphasis on the Investment Portfolio Management for Capital Asset Pricing Model. The tasks of this project are 1) review and propose linear programming approach for corporate financial planning; 2) sensitivity analysis in linear programming review and evaluation; 3) review of practical difficulties of the models’ implementation.